% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_LD.R
\name{calculate_LD}
\alias{calculate_LD}
\title{Calculate LD}
\usage{
calculate_LD(
  locus_dir,
  ld_window = 1000,
  ld_format = "r",
  plink_prefix = "plink",
  verbose = TRUE
)
}
\arguments{
\item{locus_dir}{Locus-specific results directory.}

\item{ld_window}{Set --r/--r2 max variant ct pairwise distance (usu. 10).}

\item{ld_format}{Whether to produce an LD matrix with
r (\code{ld_format="r"}) or D' (\code{ld_format="D"}) as the
pairwise SNP correlation metric.}
}
\description{
Calculate a pairwise LD matrix from a vcf file using \emph{plink}.
}
\seealso{
Other LD: 
\code{\link{LD_1KG_download_vcf}()},
\code{\link{LD_1KG}()},
\code{\link{LD_blocks}()},
\code{\link{LD_ukbiobank}()},
\code{\link{construct_subset_vcf_name}()},
\code{\link{custom_panel}()},
\code{\link{dprime_table}()},
\code{\link{filter_LD}()},
\code{\link{filter_vcf_gaston}()},
\code{\link{filter_vcf}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{index_vcf}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{leadSNP_block}()},
\code{\link{load_or_create}()},
\code{\link{plink_LD}()},
\code{\link{plink_file}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{query_vcf}()},
\code{\link{rds_to_npz}()},
\code{\link{read_bin}()},
\code{\link{read_ld_table}()},
\code{\link{run_plink_LD}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_LD}()},
\code{\link{snpstats_get_MAF}()},
\code{\link{translate_population}()},
\code{\link{vcf_to_bed}()}
}
\concept{LD}
\keyword{internal}
