% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_LD.R
\name{compute_LD}
\alias{compute_LD}
\title{Get LD using \pkg{snpStats} package}
\source{
\href{https://www.bioconductor.org/packages/release/bioc/html/snpStats.html}{snpStats Bioconductor page}
\href{https://www.bioconductor.org/packages/release/bioc/vignettes/snpStats/inst/doc/ld-vignette.pdf}{LD tutorial}
}
\usage{
compute_LD(
  ss,
  select_snps = NULL,
  stats = c("R"),
  symmetric = TRUE,
  depth = "max",
  as_sparse = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{stats}{
    A character vector specifying the linkage disequilibrium measures to
    be calculated. This should contain one or more of the strings: \code{"LLR"},
    \code{"OR"}, \code{"Q"}, \code{"Covar"}, \code{"D.prime"},
    \code{"R.squared"}, ad \code{"R"}
  }

\item{symmetric}{
    When no \code{y} argument is supplied this argument controls the format of
    the output band matrices. If \code{TRUE}, symmetric matrices are
    returned and, otherwise, an upper triangular matrices are returned
  }

\item{depth}{
    When \code{y} is not supplied, this parameter is mandatory and
    controls the maximum lag between columns of \code{x}
    considered. Thus, LD statistics are calculated between \code{x[,i]}
    and \code{x[,j]} only if \code{i} and \code{j} differ by no more
    than \code{depth}
  }

\item{LD_folder}{Locus-specific LD output folder.}
}
\description{
Get LD using \pkg{snpStats} package
}
\seealso{
Other LD: 
\code{\link{check_population_1kg}()},
\code{\link{filter_LD}()},
\code{\link{get_LD_1KG_download_vcf}()},
\code{\link{get_LD_1KG}()},
\code{\link{get_LD_UKB}()},
\code{\link{get_LD_custom}()},
\code{\link{get_LD}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{rds_to_npz}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_MAF}()}
}
\concept{LD}
\keyword{internal}
