% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_LD_blocks.R
\name{compute_LD_blocks}
\alias{compute_LD_blocks}
\title{Compute LD blocks}
\source{
\href{https://github.com/pneuvial/adjclust}{adjclust GitHub}
}
\usage{
compute_LD_blocks(
  x,
  stats = c("R.squared", "D.prime"),
  type = c("capushe", "bstick"),
  k.max = NULL,
  pct = 0.15,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{Print messages.}
}
\description{
Compute LD blocks using \link[adjclust]{snpClust}.
}
\keyword{internal}
