% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_subset_vcf_name.R
\name{construct_subset_vcf_name}
\alias{construct_subset_vcf_name}
\title{Construct the path to vcf subset}
\usage{
construct_subset_vcf_name(
  dat,
  LD_reference = NULL,
  locus_dir,
  whole_vcf = FALSE
)
}
\description{
\code{
data("locus_dir")
data("BST1")
vcf_subset <- construct_subset_vcf_name(
    dat = BST1,
    locus_dir = locus_dir,
    LD_reference = "1KGlocal"
)
}
}
\seealso{
Other LD: 
\code{\link{LD_1KG_download_vcf}()},
\code{\link{LD_1KG}()},
\code{\link{LD_blocks_cli}()},
\code{\link{LD_custom_cli}()},
\code{\link{LD_custom}()},
\code{\link{LD_ukbiobank}()},
\code{\link{calculate_LD}()},
\code{\link{compute_LD}()},
\code{\link{dprime_table}()},
\code{\link{filter_LD}()},
\code{\link{filter_vcf_cli}()},
\code{\link{filter_vcf_gaston}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{index_vcf_cli}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{leadsnp_block_plink}()},
\code{\link{load_or_create}()},
\code{\link{plink_LD}()},
\code{\link{plink_file}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{query_vcf_cli}()},
\code{\link{query_vcf}()},
\code{\link{rds_to_npz}()},
\code{\link{read_bin}()},
\code{\link{read_ld_table}()},
\code{\link{run_plink_LD}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_MAF}()},
\code{\link{translate_population}()},
\code{\link{vcf_to_bed_cli}()}
}
\concept{LD}
\keyword{internal}
