% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_panel.R
\name{custom_panel}
\alias{custom_panel}
\title{Compute LD from user-supplied VCF file}
\usage{
custom_panel(
  LD_reference,
  fullSS_genome_build = "hg19",
  LD_genome_build = "hg19",
  dat,
  locus_dir,
  force_new_LD = FALSE,
  min_r2 = FALSE,
  remove_correlates = FALSE,
  fillNA = 0,
  LD_block_size = NULL,
  remove_tmps = TRUE,
  as_sparse = TRUE,
  nThread = 1,
  conda_env = "echoR",
  verbose = TRUE
)
}
\arguments{
\item{LD_block_size}{Block size. Passed to
"--blocks-inform-frac" argument in \code{plink}. 
Recommended default value is \code{0.7}.}

\item{conda_env}{Conda environment name.}

\item{verbose}{Print messages.}
}
\description{
Compute LD from user-supplied VCF file
}
\examples{
\dontrun{ 
data("BST1")
data("locus_dir")
LD_reference <- "~/Desktop/results/Reference/custom_panel_chr4.vcf"
LD_matrix <- custom_panel(LD_reference = LD_reference, 
                          dat = BST1, 
                          locus_dir = locus_dir)
}
}
\seealso{
Other LD: 
\code{\link{LD_1KG_download_vcf}()},
\code{\link{LD_1KG}()},
\code{\link{LD_blocks}()},
\code{\link{LD_ukbiobank}()},
\code{\link{calculate_LD}()},
\code{\link{construct_subset_vcf_name}()},
\code{\link{dprime_table}()},
\code{\link{filter_LD}()},
\code{\link{filter_vcf_gaston}()},
\code{\link{filter_vcf}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{index_vcf}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{leadSNP_block}()},
\code{\link{load_or_create}()},
\code{\link{plink_LD}()},
\code{\link{plink_file}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{query_vcf}()},
\code{\link{rds_to_npz}()},
\code{\link{read_bin}()},
\code{\link{read_ld_table}()},
\code{\link{run_plink_LD}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_LD}()},
\code{\link{snpstats_get_MAF}()},
\code{\link{translate_population}()},
\code{\link{vcf_to_bed}()}
}
\concept{LD}
\keyword{internal}
