% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_UKB_LD.R
\name{download_UKB_LD}
\alias{download_UKB_LD}
\title{Download UKB LD}
\usage{
download_UKB_LD(
  LD.prefixes,
  locus_dir,
  bucket = "s3://broad-alkesgroup-ukbb-ld/",
  background = TRUE,
  force_overwrite = FALSE,
  nThread = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{locus_dir}{Storage directory to use.}

\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{force_overwrite}{Overwrite existing file.}

\item{nThread}{Number of threads to parallelize over.}

\item{verbose}{Print messages.}

\item{...}{
  Arguments passed on to \code{\link[downloadR:aws]{downloadR::aws}}
  \describe{
    \item{\code{input_url}}{URL to remote file.}
    \item{\code{output_path}}{The file name you want to save the download as.}
  }}
}
\description{
Download UK Biobank Linkage Disequilibrium, 
pre-computed in 3Mb windows by the Alkes Price lab (Broad Institute).
}
\keyword{internal}
