% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_NA.R
\name{fill_NA}
\alias{fill_NA}
\title{Fill NAs in an LD matrix}
\usage{
fill_NA(LD_matrix, fillNA = 0, verbose = FALSE)
}
\description{
Trickier than it looks.
}
\examples{
\dontrun{
data("LD_matrix")
LD_matrix <- fill_NA(LD_matrix)
}
}
