% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_vcf.R
\name{filter_vcf}
\alias{filter_vcf}
\title{Filter a vcf by min/max coordinates}
\usage{
filter_vcf(
  vcf_subset,
  popDat,
  superpopulation,
  remove_tmp = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{vcf_subset}{Path to the locus subset vcf.}

\item{popDat}{The metadata file listing the \code{superpopulation}
to which each sample belongs.}

\item{superpopulation}{Superpopulation to subset.}
}
\description{
Uses \emph{bcftools} to filter a vcf by min/max genomic coordinates
(in basepairs).
}
\seealso{
Other LD: 
\code{\link{LD_1KG_download_vcf}()},
\code{\link{LD_1KG}()},
\code{\link{LD_blocks}()},
\code{\link{LD_ukbiobank}()},
\code{\link{calculate_LD}()},
\code{\link{construct_subset_vcf_name}()},
\code{\link{custom_panel}()},
\code{\link{dprime_table}()},
\code{\link{filter_LD}()},
\code{\link{filter_vcf_gaston}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{index_vcf}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{leadSNP_block}()},
\code{\link{load_or_create}()},
\code{\link{plink_LD}()},
\code{\link{plink_file}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{query_vcf}()},
\code{\link{rds_to_npz}()},
\code{\link{read_bin}()},
\code{\link{read_ld_table}()},
\code{\link{run_plink_LD}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_LD}()},
\code{\link{snpstats_get_MAF}()},
\code{\link{translate_population}()},
\code{\link{vcf_to_bed}()}
}
\concept{LD}
\keyword{internal}
