% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_vcf_cli.R
\name{filter_vcf_cli}
\alias{filter_vcf_cli}
\title{Filter a vcf by min/max coordinates}
\usage{
filter_vcf_cli(
  vcf_subset,
  superpopulation,
  LD_reference,
  remove_tmp = TRUE,
  conda_env = "echoR",
  verbose = TRUE
)
}
\arguments{
\item{vcf_subset}{Path to the locus subset vcf.}

\item{superpopulation}{Superpopulation to subset.}
}
\description{
Uses \emph{bcftools} to filter a vcf by min/max genomic coordinates
(in basepairs).
}
\seealso{
Other LD: 
\code{\link{LD_1KG_download_vcf}()},
\code{\link{LD_1KG}()},
\code{\link{LD_blocks_cli}()},
\code{\link{LD_custom_cli}()},
\code{\link{LD_custom}()},
\code{\link{LD_ukbiobank}()},
\code{\link{calculate_LD}()},
\code{\link{compute_LD}()},
\code{\link{construct_subset_vcf_name}()},
\code{\link{dprime_table}()},
\code{\link{filter_LD}()},
\code{\link{filter_vcf_gaston}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{index_vcf_cli}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{leadsnp_block_plink}()},
\code{\link{load_or_create}()},
\code{\link{plink_LD}()},
\code{\link{plink_file}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{query_vcf_cli}()},
\code{\link{query_vcf}()},
\code{\link{rds_to_npz}()},
\code{\link{read_bin}()},
\code{\link{read_ld_table}()},
\code{\link{run_plink_LD}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_MAF}()},
\code{\link{translate_population}()},
\code{\link{vcf_to_bed_cli}()}
}
\concept{LD}
\keyword{internal}
