% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LD.R
\name{get_LD}
\alias{get_LD}
\title{Procure an LD matrix for fine-mapping}
\usage{
get_LD(
  query_dat,
  locus_dir = tempdir(),
  force_new_LD = FALSE,
  LD_reference = c("1KGphase1", "1KGphase3", "UKB"),
  target_genome = "hg19",
  samples = NULL,
  superpopulation = NULL,
  local_storage = NULL,
  leadSNP_LD_block = FALSE,
  fillNA = 0,
  verbose = TRUE,
  remove_tmps = TRUE,
  as_sparse = TRUE,
  download_method = "axel",
  conda_env = "echoR",
  nThread = 1
)
}
\arguments{
\item{query_dat}{SNP-level summary statistics subset 
to query the LD panel with.}

\item{locus_dir}{Storage directory to use.}

\item{force_new_LD}{If LD file exists, create a new one.}

\item{LD_reference}{LD reference to use:
\itemize{
\item{"1KGphase1" : }{1000 Genomes Project Phase 1}
\item{"1KGphase3" : }{1000 Genomes Project Phase 3}
\item{"UKB" : }{Pre-computed LD from a British
European-decent subset of UK Biobank.}
}}

\item{target_genome}{Genome build of the LD panel
(used only if providing custom LD panel).}

\item{samples}{[Optional] Sample names to subset the VCF by. 
If this option is used, the \link[GenomicRanges]{GRanges} object will be
converted to a \link[VariantAnnotation]{ScanVcfParam} for usage by 
\link[VariantAnnotation]{readVcf}.}

\item{superpopulation}{Superpopulation to subset LD panel by
(used only if \code{LD_reference} is "1KGphase1" or "1KGphase3".)}

\item{local_storage}{Storage folder for previously downloaded LD files.
If \code{LD_reference} is "1KGphase1" or "1KGphase3",
\code{local_storage} is where VCF files are stored.
If \code{LD_reference} is "UKB", \code{local_storage} is where
LD compressed numpy array (npz) files are stored.
Set to \code{NULL} to download VCFs/LD npz from remote storage system.}

\item{leadSNP_LD_block}{Only return SNPs within the same LD block
as the lead SNP (the SNP with the smallest p-value).}

\item{fillNA}{Value to fill LD matrix NAs with.}

\item{verbose}{Print messages.}

\item{remove_tmps}{Remove all intermediate files
like VCF, npz, and plink files.}

\item{as_sparse}{Convert the LD matrix to a sparse matrix.}

\item{download_method}{\itemize{
\item{\code{"axel"} : }{Multi-threaded}
\item{\code{"wget"} : }{Single-threaded}
\item{\code{"download.file"} : }{Single-threaded}
\item{\code{"internal"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
\item{\code{"wininet"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
\item{\code{"libcurl"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
\item{\code{"curl"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
}
 or
\code{"download.file"} (single-threaded) .}

\item{conda_env}{Conda environment to use.}

\item{nThread}{Number of threads to parallelize over.}
}
\value{
A named list containing:
\itemize{
\item{"LD": }{Symmetric LD matrix of pairwise SNP correlations.}
\item{"DT": }{Standardised query data filtered to only the 
SNPs included in both \code{query_dat} and the LD matrix.}
\item{"RDS_path": }{The path to where the LD matrix was saved.}
}
}
\description{
Calculate and/or query linkage disequilibrium (LD) from reference panels
 (UK Biobank, 1000 Genomes), a user-supplied pre-computed LD matrix
}
\details{
Options:
\itemize{
\item{Download pre-computed LD matrix from UK Biobank.}
\item{Download raw VCF file from 1KG and compute LD on the fly.}
\item{Compute LD on the fly from a user-supplied VCF file.}
\item{Use a user-supplied pre-computed LD-matrix.}
}
}
\examples{
query_dat <- echodata::BST1[seq(1, 50), ] 
locus_dir <- file.path(tempdir(), echodata::locus_dir) 

LD_list <- echoLD::get_LD(
    locus_dir = locus_dir,
    query_dat = query_dat,
    LD_reference = "1KGphase1")
}
\seealso{
Other LD: 
\code{\link{check_population_1kg}()},
\code{\link{compute_LD}()},
\code{\link{filter_LD}()},
\code{\link{get_LD_1KG_download_vcf}()},
\code{\link{get_LD_1KG}()},
\code{\link{get_LD_UKB}()},
\code{\link{get_LD_custom}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{rds_to_npz}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_MAF}()}
}
\concept{LD}
