% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD_1KG_download_vcf.R
\name{get_LD_1KG_download_vcf}
\alias{get_LD_1KG_download_vcf}
\title{Download VCF subset from 1000 Genomes}
\source{
\code{
query_dat <- echodata::BST1
locus_dir <- file.path(tempdir(), echodata::locus_dir)
query_granges <- echotabix::construct_query(query_dat=query_dat)

vcf_subset.popDat <- echoLD:::get_LD_1KG_download_vcf(
    query_granges = query_granges,
    LD_reference = "1KGphase1",
    locus_dir = locus_dir)
}
}
\usage{
get_LD_1KG_download_vcf(
  query_granges,
  query_genome = "hg19",
  LD_reference = "1KGphase1",
  superpopulation = NULL,
  samples = character(0),
  local_storage = NULL,
  locus_dir = tempdir(),
  save_path = echotabix::construct_vcf_path(locus_dir = locus_dir, subdir = "LD",
    target_path = LD_reference, query_granges = query_granges),
  query_save = TRUE,
  force_new = FALSE,
  conda_env = "echoR_mini",
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{query_granges}{\link[GenomicRanges]{GRanges} object 
to be used for querying the \code{target_path} file. 
Alternatively, can be variant-level summary statistics to be converted into a
\link[GenomicRanges]{GRanges} object by \link[echotabix]{construct_query}.}

\item{query_genome}{Genome build of the \code{query_dat}.}

\item{LD_reference}{LD reference to use:
\itemize{
\item{"1KGphase1" : }{1000 Genomes Project Phase 1 (genome build: hg19).}
\item{"1KGphase3" : }{1000 Genomes Project Phase 3 (genome build: hg19).}
\item{"UKB" : }{Pre-computed LD from a British
European-decent subset of UK Biobank (genome build: hg19).}
\item{"<vcf_path>" : }{User-supplied path to a custom VCF file 
to compute LD matrix from 
(genome build: defined by user with \code{target_genome}).}
}}

\item{superpopulation}{Superpopulation to subset LD panel by
(used only if \code{LD_reference} is "1KGphase1" or "1KGphase3").
See \link[echoLD]{popDat_1KGphase1} and \link[echoLD]{popDat_1KGphase3}
for full tables of their respective samples.}

\item{samples}{[Optional] Sample names to subset the VCF by. 
If this option is used, the \link[GenomicRanges]{GRanges} object will be
converted to a \link[VariantAnnotation]{ScanVcfParam} for usage by 
\link[VariantAnnotation]{readVcf}.}

\item{local_storage}{Storage folder for previously downloaded LD files.
If \code{LD_reference} is "1KGphase1" or "1KGphase3",
\code{local_storage} is where VCF files are stored.
If \code{LD_reference} is "UKB", \code{local_storage} is where
LD compressed numpy array (npz) files are stored.
Set to \code{NULL} to download VCFs/LD npz from remote storage system.}

\item{locus_dir}{Storage directory to use.}

\item{save_path}{Path to save LD subset to.}

\item{query_save}{Whether to save the queried data subset.}

\item{conda_env}{Conda environments to search in. 
If \code{NULL} (default), will search all conda environments.}

\item{nThread}{Number of threads to parallelize over.}

\item{verbose}{Print messages.}
}
\description{
Query the 1000 Genomes Project for a subset of their individual-level VCF
files.
}
\seealso{
Other LD: 
\code{\link{check_population_1kg}()},
\code{\link{compute_LD}()},
\code{\link{filter_LD}()},
\code{\link{get_LD_1KG}()},
\code{\link{get_LD_UKB}()},
\code{\link{get_LD_custom}()},
\code{\link{get_LD}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{rds_to_npz}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_MAF}()}
}
\concept{LD}
\keyword{internal}
