% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LD_blocks.R
\name{get_LD_blocks}
\alias{get_LD_blocks}
\title{Get LD blocks}
\source{
\href{https://github.com/pneuvial/adjclust}{adjclust GitHub}
}
\usage{
get_LD_blocks(
  query_dat,
  ss,
  stats = c("R.squared", "D.prime"),
  pct = 0.15,
  verbose = TRUE
)
}
\arguments{
\item{query_dat}{SNP-level data table.}

\item{ss}{\pkg{snpStats} object or LD matrix containing 
\ifelse{html}{\out{r}}{\eqn{r}} or 
\ifelse{html}{\out{r<sup>2</sup>}}{\eqn{r^2}} values.}

\item{stats}{Character vector of LD statistics to compute.
Passed to \code{\link[adjclust]{snpClust}}.}

\item{pct}{Numeric. Percentage of the maximum number of clusters
to select. Passed to \code{\link[adjclust]{select}}.}

\item{verbose}{Print messages.}
}
\value{
A list with the input data and LD matrix 
(\ifelse{html}{\out{r<sup>2</sup>}}{\eqn{r^2}}).
}
\description{
Identify the LD block in which the lead SNP resides.
}
