% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LD_custom.R
\name{get_LD_custom}
\alias{get_LD_custom}
\title{Compute LD from 1000 Genomes}
\usage{
get_LD_custom(
  locus_dir = tempdir(),
  query_dat,
  LD_reference,
  target_genome = "GRCh37",
  superpopulation = NULL,
  samples = NULL,
  local_storage = NULL,
  leadSNP_LD_block = FALSE,
  force_new = FALSE,
  force_new_MAF = FALSE,
  fillNA = 0,
  stats = "R",
  as_sparse = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{locus_dir}{Storage directory to use.}

\item{query_dat}{SNP-level summary statistics subset 
to query the LD panel with.}

\item{LD_reference}{LD reference to use:
\itemize{
\item{"1KGphase1" : }{1000 Genomes Project Phase 1}
\item{"1KGphase3" : }{1000 Genomes Project Phase 3}
\item{"UKB" : }{Pre-computed LD from a British
European-decent subset of UK Biobank.}
}}

\item{target_genome}{Genome build of the LD panel
(used only if providing custom LD panel).}

\item{superpopulation}{Superpopulation to subset LD panel by
(used only if \code{LD_reference} is "1KGphase1" or "1KGphase3".)}

\item{samples}{[Optional] Sample names to subset the VCF by. 
If this option is used, the \link[GenomicRanges]{GRanges} object will be
converted to a \link[VariantAnnotation]{ScanVcfParam} for usage by 
\link[VariantAnnotation]{readVcf}.}

\item{local_storage}{Storage folder for previously downloaded LD files.
If \code{LD_reference} is "1KGphase1" or "1KGphase3",
\code{local_storage} is where VCF files are stored.
If \code{LD_reference} is "UKB", \code{local_storage} is where
LD compressed numpy array (npz) files are stored.
Set to \code{NULL} to download VCFs/LD npz from remote storage system.}

\item{leadSNP_LD_block}{Only return SNPs within the same LD block
as the lead SNP (the SNP with the smallest p-value).}

\item{force_new}{Force the creation of a new VCF subset file
even if one exists.}

\item{fillNA}{When pairwise LD (r) between two SNPs is \code{NA},
replace with 0.}

\item{as_sparse}{Convert the LD matrix to a sparse matrix.}

\item{verbose}{Print messages.}
}
\description{
Downloads a subset vcf of the 1KG database that matches
 your locus coordinates. Then uses \link[snpStats]{ld}
  to calculate LD on the fly.
}
\details{
This approach is taken, because other API query tools have
limitations with the window size being queried.
This approach does not have this limitations,
allowing you to fine-map loci more completely.
}
\examples{
\dontrun{
query_dat <-  echodata::BST1[seq(1, 50), ]
locus_dir <- echodata::locus_dir
locus_dir <- file.path(tempdir(), locus_dir) 
LD_reference <- system.file("extdata", "BST1.1KGphase3.vcf.bgz",
    package = "echoLD"
)
LD_list <- get_LD_custom(
    locus_dir = locus_dir,
    query_dat = query_dat
    LD_reference = LD_reference
)
}
}
\seealso{
Other LD: 
\code{\link{check_population_1kg}()},
\code{\link{compute_LD}()},
\code{\link{filter_LD}()},
\code{\link{get_LD_1KG_download_vcf}()},
\code{\link{get_LD_1KG}()},
\code{\link{get_LD_UKB}()},
\code{\link{get_LD}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{rds_to_npz}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_MAF}()}
}
\concept{LD}
\keyword{internal}
