% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LD_matrix.R
\name{get_LD_matrix}
\alias{get_LD_matrix}
\title{Get LD from pre-computed matrix}
\usage{
get_LD_matrix(
  locus_dir = tempdir(),
  query_dat,
  LD_reference,
  query_genome = "hg19",
  target_genome = "hg19",
  fillNA = 0,
  stats = "R",
  as_sparse = TRUE,
  subset_common = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{locus_dir}{Storage directory to use.}

\item{query_dat}{SNP-level summary statistics subset 
to query the LD panel with.}

\item{LD_reference}{LD reference to use:
\itemize{
\item{"1KGphase1" : }{1000 Genomes Project Phase 1 (genome build: hg19).}
\item{"1KGphase3" : }{1000 Genomes Project Phase 3 (genome build: hg19).}
\item{"UKB" : }{Pre-computed LD from a British
European-decent subset of UK Biobank.
\emph{Genome build} : hg19}
\item{"<vcf_path>" : }{User-supplied path to a custom VCF file 
to compute LD matrix from.\cr
\emph{Accepted formats}: \emph{.vcf} / \emph{.vcf.gz} / \emph{.vcf.bgz}\cr
\emph{Genome build} : defined by user with \code{target_genome}.}
\item{"<matrix_path>" : }{User-supplied path to a pre-computed LD matrix   
\emph{Accepted formats}:  \emph{.rds} / \emph{.rda} / \emph{.csv} / 
\emph{.tsv} / \emph{.txt}\cr
\emph{Genome build} : defined by user with \code{target_genome}.}
}}

\item{query_genome}{Genome build of the \code{query_dat}.}

\item{target_genome}{Genome build of the LD panel. 
This is automatically assigned to the correct genome build for each LD panel
 except when the user supplies custom vcf/LD files.}

\item{fillNA}{When pairwise LD between two SNPs is \code{NA},
replace with 0.}

\item{stats}{The linkage disequilibrium measure to be calculated. 
Can be \code{"R"} (default) or \code{"R.squared"}.}

\item{as_sparse}{Convert the LD matrix to a sparse matrix.}

\item{subset_common}{Subset \code{LD_matrix} and \code{dat} to only the 
SNPs that are common to them both.}

\item{verbose}{Print messages.}
}
\description{
Get LD from a pre-computed matrix of pairwise 
\ifelse{html}{\out{r}}{\eqn{r}} / 
\ifelse{html}{\out{r<sup>2</sup>}}{\eqn{r^2}}values.
}
\examples{
query_dat <-  echodata::BST1[seq(1, 50), ] 
locus_dir <- file.path(tempdir(),  echodata::locus_dir)
LD_reference <- tempfile(fileext = ".csv")
utils::write.csv(echodata::BST1_LD_matrix,
                 file = LD_reference,
                 row.names = TRUE)
LD_list <- get_LD_matrix(
    locus_dir = locus_dir,
    query_dat = query_dat,
    LD_reference = LD_reference)
}
\seealso{
Other LD: 
\code{\link{check_population_1kg}()},
\code{\link{compute_LD}()},
\code{\link{filter_LD}()},
\code{\link{get_LD}()},
\code{\link{get_LD_1KG}()},
\code{\link{get_LD_1KG_download_vcf}()},
\code{\link{get_LD_UKB}()},
\code{\link{get_LD_vcf}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{rds_to_npz}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_MAF}()}
}
\concept{LD}
