% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LD_UKB.R
\name{get_LD_UKB}
\alias{get_LD_UKB}
\title{Download LD matrices from UK Biobank}
\usage{
get_LD_UKB(
  query_dat,
  query_genome = "hg19",
  locus_dir,
  chrom = NULL,
  min_pos = NULL,
  force_new_LD = FALSE,
  local_storage = NULL,
  download_full_ld = FALSE,
  download_method = "axel",
  fillNA = 0,
  nThread = 1,
  return_matrix = TRUE,
  as_sparse = TRUE,
  conda_env = "echoR_mini",
  remove_tmps = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{query_dat}{SNP-level summary statistics subset 
to query the LD panel with.}

\item{query_genome}{Genome build of the \code{query_dat}.}

\item{locus_dir}{Storage directory to use.}

\item{force_new_LD}{Force new LD subset.}

\item{local_storage}{Path to folder with previously download LD npz files.}

\item{download_method}{If "python" will import compressed numpy array
directly into R using \pkg{reticulate}. Otherwise, will be passed to
\link[downloadR]{downloader} to download the full 3Mb-window LD matrix first.}

\item{fillNA}{Value to fill LD matrix NAs with.}

\item{nThread}{Number of threads to parallelize over.}

\item{as_sparse}{Convert the LD matrix to a sparse matrix.}

\item{conda_env}{Conda environments to search in. 
If \code{NULL} (default), will search all conda environments.}

\item{remove_tmps}{Remove all intermediate files
like VCF, npz, and plink files.}

\item{verbose}{Print messages.}
}
\description{
Download pre-computed LD matrices from
\href{https://www.ukbiobank.ac.uk}{UK Biobank} in 3Mb windows,
then subset to the region that overlaps with \code{query_dat}.
}
\details{
LD was derived from a  British, European-decent subpopulation
in the UK Biobank. LD was pre-computed and stored by the Alkes Price lab
(see \href{https://www.biorxiv.org/content/10.1101/807792v3}{here}).
}
\seealso{
Other LD: 
\code{\link{check_population_1kg}()},
\code{\link{compute_LD}()},
\code{\link{filter_LD}()},
\code{\link{get_LD_1KG_download_vcf}()},
\code{\link{get_LD_1KG}()},
\code{\link{get_LD_custom}()},
\code{\link{get_LD}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{rds_to_npz}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_MAF}()}
}
\concept{LD}
\keyword{internal}
