% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lead_r2.R
\name{get_lead_r2}
\alias{get_lead_r2}
\title{Find correlates of the lead GWAS/QTL SNP}
\usage{
get_lead_r2(
  finemap_dat,
  LD_matrix = NULL,
  fillNA = 0,
  LD_format = "matrix",
  verbose = TRUE
)
}
\description{
Find correlates of the lead GWAS/QTL SNP
}
