% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locus_vcf_folder.R
\name{get_locus_vcf_folder}
\alias{get_locus_vcf_folder}
\title{Get VCF storage folder}
\usage{
get_locus_vcf_folder(locus_dir = NULL)
}
\description{
Get VCF storage folder
}
\seealso{
Other LD: 
\code{\link{check_population_1kg}()},
\code{\link{compute_LD}()},
\code{\link{filter_LD}()},
\code{\link{get_LD_1KG_download_vcf}()},
\code{\link{get_LD_1KG}()},
\code{\link{get_LD_UKB}()},
\code{\link{get_LD_matrix}()},
\code{\link{get_LD_vcf}()},
\code{\link{get_LD}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{rds_to_npz}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_MAF}()}
}
\concept{LD}
\keyword{internal}
