% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.1KG_download_vcf.R
\name{LD_1KG_download_vcf}
\alias{LD_1KG_download_vcf}
\title{Download vcf subset from 1000 Genomes}
\usage{
LD_1KG_download_vcf(
  dat,
  LD_reference = "1KGphase1",
  remote_LD = TRUE,
  vcf_folder = NULL,
  locus_dir,
  locus = NULL,
  whole_vcf = FALSE,
  download_method = "wget",
  force_new_vcf = FALSE,
  query_by_regions = FALSE,
  remove_tmps = TRUE,
  nThread = 1,
  conda_env = "echoR",
  verbose = TRUE
)
}
\arguments{
\item{LD_reference}{Which linkage disequilibrium reference panel do you want to use.
Options include:
\describe{
\item{"UKB"}{A pre-caclulated LD reference matrix from a subset of caucasian British individuals from the UK Biobank. See \href{https://www.biorxiv.org/content/10.1101/807792v2}{Wiessbrod et al. (2019)} for more details.}
\item{"1KGphase1"}{Download a subset of the 1000 Genomes Project Phase 1 vcf and calculate LD on the fly with plink.}
\item{"1KGphase3"}{Download a subset of the 1000 Genomes Project Phase 3 vcf and calculate LD on the fly with plink.}
\item{"<path>/*.vcf" or "<path>/*.vcf.gz"}{Alternatively, users can provide their own custom panel by supplying a list of \emph{.vcf} file path (one per locus) which \pkg{echolocatoR} will use to compute LD (using \emph{plink}).}
}}

\item{remote_LD}{When acquiring LD matrices,
the default is to delete the full vcf or npz files after \pkg{echolocatoR} has extracted the necssary subset.
However, if you wish to keep these full files (which can be quite large) set \code{remote_LD=T}.}

\item{query_by_regions}{You can make queries with \code{tabix}
in two different ways:

\describe{
\item{\code{query_by_regions=F} \emph{(default)}}{
Return a vcf with all positions between the min/max in \code{dat}.
Takes up more storage but is MUCH faster}
\item{\code{query_by_regions=T}}{
Return a vcf with only the exact positions present in \code{dat}.
 Takes up less storage but is MUCH slower}
}}

\item{remove_tmps}{Whether to remove any temporary files (e.g. FINEMAP output files) after the pipeline is done running.}

\item{conda_env}{The name of a conda environment to use.}

\item{verbose}{Whether \pkg{echolocatoR} should be verbose or silent.}
}
\description{
Download vcf subset from 1000 Genomes
}
\examples{
\dontrun{
data("BST1")
vcf_subset.popDat <- LD_1KG_download_vcf(
    dat = BST1,
    LD_reference = "1KGphase1",
    locus_dir = file.path(tempdir(), locus_dir)
)
}
}
\seealso{
Other LD: 
\code{\link{LD_1KG}()},
\code{\link{LD_blocks}()},
\code{\link{LD_ukbiobank}()},
\code{\link{calculate_LD}()},
\code{\link{construct_subset_vcf_name}()},
\code{\link{custom_panel}()},
\code{\link{dprime_table}()},
\code{\link{filter_LD}()},
\code{\link{filter_vcf_gaston}()},
\code{\link{filter_vcf}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{index_vcf}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{leadSNP_block}()},
\code{\link{load_or_create}()},
\code{\link{plink_LD}()},
\code{\link{plink_file}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{query_vcf}()},
\code{\link{rds_to_npz}()},
\code{\link{read_bin}()},
\code{\link{read_ld_table}()},
\code{\link{run_plink_LD}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_LD}()},
\code{\link{snpstats_get_MAF}()},
\code{\link{translate_population}()},
\code{\link{vcf_to_bed}()}
}
\concept{LD}
\keyword{internal}
