% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD_blocks.R
\name{LD_blocks}
\alias{LD_blocks}
\title{Calculate LD blocks.}
\usage{
LD_blocks(
  bed_bim_fam,
  LD_block_size = 0.7,
  conda_env = "echoR",
  verbose = TRUE
)
}
\arguments{
\item{LD_block_size}{Block size. Passed to
"--blocks-inform-frac" argument in \code{plink}. 
Recommended default value is \code{0.7}.}

\item{conda_env}{Conda environment name.}

\item{verbose}{Print messages.}

\item{LD_folder}{Path to save LD to.}
}
\description{
Uses \emph{plink} to group highly correlated SNPs into LD blocks.
}
\seealso{
Other LD: 
\code{\link{LD_1KG_download_vcf}()},
\code{\link{LD_1KG}()},
\code{\link{LD_ukbiobank}()},
\code{\link{calculate_LD}()},
\code{\link{construct_subset_vcf_name}()},
\code{\link{custom_panel}()},
\code{\link{dprime_table}()},
\code{\link{filter_LD}()},
\code{\link{filter_vcf_gaston}()},
\code{\link{filter_vcf}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{index_vcf}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{leadSNP_block}()},
\code{\link{load_or_create}()},
\code{\link{plink_LD}()},
\code{\link{plink_file}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{query_vcf}()},
\code{\link{rds_to_npz}()},
\code{\link{read_bin}()},
\code{\link{read_ld_table}()},
\code{\link{run_plink_LD}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_LD}()},
\code{\link{snpstats_get_MAF}()},
\code{\link{translate_population}()},
\code{\link{vcf_to_bed}()}
}
\concept{LD}
\keyword{internal}
