% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD_custom_cli.R
\name{LD_custom_cli}
\alias{LD_custom_cli}
\title{Compute LD from user-supplied VCF file}
\usage{
LD_custom_cli(
  LD_reference,
  superpopulation = NULL,
  fullSS_genome_build = "hg19",
  LD_genome_build = "hg19",
  dat,
  locus_dir,
  force_new_LD = FALSE,
  min_r2 = FALSE,
  remove_correlates = FALSE,
  fillNA = 0,
  LD_block_size = NULL,
  remove_tmps = TRUE,
  as_sparse = TRUE,
  nThread = 1,
  conda_env = "echoR",
  verbose = TRUE
)
}
\arguments{
\item{LD_block_size}{Block size. Passed to
"--blocks-inform-frac" argument in \code{plink}.
Recommended default value is \code{0.7}.}

\item{conda_env}{Conda environment name.}

\item{verbose}{Print messages.}
}
\description{
Compute LD from user-supplied VCF file
}
\examples{
\dontrun{
data("BST1")
data("locus_dir")
LD_reference <- system.file("exdata", "BST1.1KGphase3.vcf.bgz",
    package = "echoLD"
)
LD_matrix <- LD_custom_cli(
    LD_reference = LD_reference,
    dat = BST1,
    locus_dir = locus_dir
)
}
}
\seealso{
Other LD: 
\code{\link{LD_1KG_download_vcf}()},
\code{\link{LD_1KG}()},
\code{\link{LD_blocks_cli}()},
\code{\link{LD_custom}()},
\code{\link{LD_ukbiobank}()},
\code{\link{calculate_LD}()},
\code{\link{compute_LD}()},
\code{\link{construct_subset_vcf_name}()},
\code{\link{dprime_table}()},
\code{\link{filter_LD}()},
\code{\link{filter_vcf_cli}()},
\code{\link{filter_vcf_gaston}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{index_vcf_cli}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{leadsnp_block_plink}()},
\code{\link{load_or_create}()},
\code{\link{plink_LD}()},
\code{\link{plink_file}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{query_vcf_cli}()},
\code{\link{query_vcf}()},
\code{\link{rds_to_npz}()},
\code{\link{read_bin}()},
\code{\link{read_ld_table}()},
\code{\link{run_plink_LD}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_MAF}()},
\code{\link{translate_population}()},
\code{\link{vcf_to_bed_cli}()}
}
\concept{LD}
\keyword{internal}
