% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadsnp_block_plink.R
\name{leadsnp_block_plink}
\alias{leadsnp_block_plink}
\title{Identify the LD block in which the lead SNP resides}
\usage{
leadsnp_block_plink(
  leadSNP,
  bed_bim_fam,
  LD_block_size = 0.7,
  verbose = TRUE,
  conda_env = "echoR"
)
}
\arguments{
\item{LD_block_size}{Block size. Passed to
"--blocks-inform-frac" argument in \code{plink}.
Recommended default value is \code{0.7}.}

\item{verbose}{Print messages.}

\item{conda_env}{Conda environment name.}
}
\description{
Identify the LD block in which the lead SNP resides
}
\seealso{
Other LD: 
\code{\link{LD_1KG_download_vcf}()},
\code{\link{LD_1KG}()},
\code{\link{LD_blocks}()},
\code{\link{LD_custom_cli}()},
\code{\link{LD_custom}()},
\code{\link{LD_ukbiobank}()},
\code{\link{calculate_LD}()},
\code{\link{compute_LD}()},
\code{\link{construct_subset_vcf_name}()},
\code{\link{dprime_table}()},
\code{\link{filter_LD}()},
\code{\link{filter_vcf_cli}()},
\code{\link{filter_vcf_gaston}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{index_vcf_cli}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{load_or_create}()},
\code{\link{plink_LD}()},
\code{\link{plink_file}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{query_vcf_cli}()},
\code{\link{query_vcf}()},
\code{\link{rds_to_npz}()},
\code{\link{read_bin}()},
\code{\link{read_ld_table}()},
\code{\link{run_plink_LD}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_MAF}()},
\code{\link{translate_population}()},
\code{\link{vcf_to_bed_cli}()}
}
\concept{LD}
\keyword{internal}
