% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink_LD.R
\name{plink_LD}
\alias{plink_LD}
\title{Calculate LD}
\usage{
plink_LD(
  leadSNP = NULL,
  dat,
  bim_path = NULL,
  remove_excess_snps = TRUE,
  merge_by_RSID = FALSE,
  LD_folder,
  min_r2 = FALSE,
  min_Dprime = FALSE,
  remove_correlates = FALSE,
  fillNA = 0,
  plink_prefix = "plink",
  verbose = TRUE,
  conda_env = NULL
)
}
\description{
Use \emph{plink} to calculate LD from a vcf.
}
\examples{
locus_dir <- "/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Kunkle_2019/ACE"
LD_folder <- file.path(locus_dir, "LD")
ld.matrix <- plink_LD(dat = BST1, LD_folder = LD_folder)
}
\seealso{
Other LD: 
\code{\link{LD_1KG_download_vcf}()},
\code{\link{LD_1KG}()},
\code{\link{LD_blocks}()},
\code{\link{LD_ukbiobank}()},
\code{\link{calculate_LD}()},
\code{\link{construct_subset_vcf_name}()},
\code{\link{custom_panel}()},
\code{\link{dprime_table}()},
\code{\link{filter_LD}()},
\code{\link{filter_vcf_gaston}()},
\code{\link{filter_vcf}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{index_vcf}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{leadSNP_block}()},
\code{\link{load_or_create}()},
\code{\link{plink_file}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{query_vcf}()},
\code{\link{rds_to_npz}()},
\code{\link{read_bin}()},
\code{\link{read_ld_table}()},
\code{\link{run_plink_LD}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_LD}()},
\code{\link{snpstats_get_MAF}()},
\code{\link{translate_population}()},
\code{\link{vcf_to_bed}()}
}
\concept{LD}
\keyword{internal}
