% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printer.R
\name{printer}
\alias{printer}
\title{printer}
\usage{
printer(..., v = T)
}
\description{
Concatenate and print any number of items.
}
\seealso{
Other general: 
\code{\link{get_os}()},
\code{\link{tryFunc}()}
}
\concept{general}
\keyword{internal}
