% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vcf.R
\name{query_vcf}
\alias{query_vcf}
\title{Query vcf file}
\usage{
query_vcf(
  dat,
  vcf_URL,
  locus_dir,
  LD_reference,
  whole_vcf = FALSE,
  force_new_vcf = FALSE,
  remove_original_vcf = FALSE,
  download_method = "wget",
  query_by_regions = FALSE,
  nThread = 1,
  conda_env = "echoR",
  verbose = TRUE
)
}
\description{
Query vcf file
}
\examples{
\dontrun{
data("locus_dir")
data("BST1")
# Custom
LD_reference <- "~/Desktop/results/Reference/custom_panel_chr4.vcf"
vcf_file <- index_vcf(vcf_file = LD_reference)
vcf_subset <- query_vcf(
    dat = BST1,
    locus_dir = locus_dir,
    vcf_URL = vcf_file,
    LD_reference = LD_reference,
    force_new_vcf = TRUE
)
}
}
\seealso{
Other LD: 
\code{\link{LD_1KG_download_vcf}()},
\code{\link{LD_1KG}()},
\code{\link{LD_blocks}()},
\code{\link{LD_ukbiobank}()},
\code{\link{calculate_LD}()},
\code{\link{construct_subset_vcf_name}()},
\code{\link{custom_panel}()},
\code{\link{dprime_table}()},
\code{\link{filter_LD}()},
\code{\link{filter_vcf_gaston}()},
\code{\link{filter_vcf}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{index_vcf}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{leadSNP_block}()},
\code{\link{load_or_create}()},
\code{\link{plink_LD}()},
\code{\link{plink_file}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{rds_to_npz}()},
\code{\link{read_bin}()},
\code{\link{read_ld_table}()},
\code{\link{run_plink_LD}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_LD}()},
\code{\link{snpstats_get_MAF}()},
\code{\link{translate_population}()},
\code{\link{vcf_to_bed}()}
}
\concept{LD}
\keyword{internal}
