% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vcf_cli.R
\name{query_vcf_cli}
\alias{query_vcf_cli}
\title{Query VCF file: CLI}
\usage{
query_vcf_cli(
  dat,
  vcf_url,
  locus_dir,
  LD_reference,
  whole_vcf = FALSE,
  force_new_vcf = FALSE,
  remove_original_vcf = FALSE,
  download_method = "download.file",
  query_by_regions = FALSE,
  nThread = 1,
  conda_env = "echoR",
  verbose = TRUE
)
}
\description{
Query a (remote) VCF file using the Command Line Interface (CLI)
}
\seealso{
Other LD: 
\code{\link{LD_1KG_download_vcf}()},
\code{\link{LD_1KG}()},
\code{\link{LD_blocks}()},
\code{\link{LD_custom_cli}()},
\code{\link{LD_custom}()},
\code{\link{LD_ukbiobank}()},
\code{\link{calculate_LD}()},
\code{\link{compute_LD}()},
\code{\link{construct_subset_vcf_name}()},
\code{\link{dprime_table}()},
\code{\link{filter_LD}()},
\code{\link{filter_vcf_cli}()},
\code{\link{filter_vcf_gaston}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{index_vcf_cli}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{leadsnp_block_plink}()},
\code{\link{load_or_create}()},
\code{\link{plink_LD}()},
\code{\link{plink_file}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{query_vcf}()},
\code{\link{rds_to_npz}()},
\code{\link{read_bin}()},
\code{\link{read_ld_table}()},
\code{\link{run_plink_LD}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_MAF}()},
\code{\link{translate_population}()},
\code{\link{vcf_to_bed_cli}()}
}
\concept{LD}
\keyword{internal}
