% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_to_r.R
\name{r2_to_r}
\alias{r2_to_r}
\title{Convert \ifelse{html}{\out{r<sup>2</sup>}}{\eqn{r^2}}
 to \ifelse{html}{\out{r}}{\eqn{r}}, and vice versa}
\usage{
r2_to_r(LD_matrix, LD_path = "", stats = "R", verbose = TRUE)
}
\arguments{
\item{LD_matrix}{LD matrix.}

\item{LD_path}{[Optional] Path where LD matrix is stored. This is used to 
determine whether the matrix has already been converted to absolute 
 \ifelse{html}{\out{r}}{\eqn{r}} based on the presence of the 
 "rAbsolute" substring (to avoid taking the square root twice).}

\item{stats}{The linkage disequilibrium measure to be calculated. 
Can be \code{"R"} (default) or \code{"R.squared"}.}

\item{verbose}{Print messages.}
}
\value{
Named list containing the LD matrix and which LD metric 
is has been formatted to.
}
\description{
Infers which LD correlation metric the \code{LD_matrix} currently contains, 
and converts it to the requested LD correlation metric.
\bold{NOTE: } Since +/- valence cannot be recovered from an 
\ifelse{html}{\out{r<sup>2</sup>}}{\eqn{r^2}} matrix, 
\ifelse{html}{\out{r<sup>2</sup>}}{\eqn{r^2}}
 can only be converted to \emph{absolute} 
 \ifelse{html}{\out{r}}{\eqn{r}}.
}
\examples{
LD_matrix <- echodata::BST1_LD_matrix
r2r_out <- r2_to_r(LD_matrix = LD_matrix)
}
