% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_to_npz.R
\name{rds_to_npz}
\alias{rds_to_npz}
\title{Convert .RDS file back to .npz format}
\usage{
rds_to_npz(rds_path, conda_env = "echoR", verbose = TRUE)
}
\description{
Convert .RDS file back to .npz format
}
\seealso{
Other LD: 
\code{\link{LD_1KG_download_vcf}()},
\code{\link{LD_1KG}()},
\code{\link{LD_custom}()},
\code{\link{LD_ukbiobank}()},
\code{\link{compute_LD}()},
\code{\link{filter_LD}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{load_or_create}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_MAF}()},
\code{\link{translate_population}()}
}
\concept{LD}
\keyword{internal}
