% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSparse.R
\name{readSparse}
\alias{readSparse}
\title{Read LD matrix}
\usage{
readSparse(LD_path, as_sparse = TRUE, as_df = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{LD_path}{Path to LD matrix.}

\item{as_sparse}{Convert the LD matrix to a sparse matrix.}

\item{as_df}{Convert to the matrix to a \link[base]{data.frame}.}

\item{verbose}{Print messages.}

\item{...}{
  Arguments passed on to \code{\link[downloadR:load_rdata]{downloadR::load_rdata}}
  \describe{
    \item{\code{fileName}}{Name of the file to load.}
  }}
}
\description{
Read in an LD matrix stored in a variety of file formats, and convert it
to a sprase matrix.
}
\examples{
LD_path <- tempfile(fileext = ".csv")
utils::write.csv(echodata::BST1_LD_matrix,
                 file = LD_path,
                 row.names = TRUE)
ld_mat <- readSparse(LD_path = LD_path)
}
