% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveSparse.R
\name{saveSparse}
\alias{saveSparse}
\title{Save LD matrix as a sparse matrix}
\usage{
saveSparse(LD_matrix, LD_path = tempfile(fileext = ".rds"), verbose = TRUE)
}
\arguments{
\item{LD_matrix}{LD matrix to save.}

\item{LD_path}{Path to save LD matrix to.}

\item{verbose}{Print messages.}
}
\description{
Converting LD matrices to sparse format reduces file size by half.
}
\examples{
echodata::BST1_LD_matrix
LD_path <- saveSparse(LD_matrix = LD_matrix)
}
\seealso{
Other LD: 
\code{\link{check_population_1kg}()},
\code{\link{compute_LD}()},
\code{\link{filter_LD}()},
\code{\link{get_LD_1KG_download_vcf}()},
\code{\link{get_LD_1KG}()},
\code{\link{get_LD_UKB}()},
\code{\link{get_LD_matrix}()},
\code{\link{get_LD_vcf}()},
\code{\link{get_LD}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{rds_to_npz}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_MAF}()}
}
\concept{LD}
