% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_vcf_samples.R
\name{select_vcf_samples}
\alias{select_vcf_samples}
\title{Subset VCF samples}
\usage{
select_vcf_samples(
  superpopulation = NULL,
  samples = NULL,
  LD_reference = "1KGphase1",
  verbose = TRUE
)
}
\arguments{
\item{samples}{Sample names to subset the VCF by before computing LD.}

\item{verbose}{Print messages.}
}
\description{
Subset VCF samples
}
\keyword{internal}
