% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpstats_get_LD.R
\name{snpstats_get_LD}
\alias{snpstats_get_LD}
\title{Get LD using \pkg{snpStats} package}
\source{
\href{https://www.bioconductor.org/packages/release/bioc/html/snpStats.html}{snpStats Bioconductor page}
\href{https://www.bioconductor.org/packages/release/bioc/vignettes/snpStats/inst/doc/ld-vignette.pdf}{LD tutorial}
}
\usage{
snpstats_get_LD(
  bed_bim_fam,
  select_snps = NULL,
  stats = c("R"),
  symmetric = TRUE,
  depth = "max",
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{stats}{
    A character vector specifying the linkage disequilibrium measures to
    be calculated. This should contain one or more of the strings: \code{"LLR"},
    \code{"OR"}, \code{"Q"}, \code{"Covar"}, \code{"D.prime"},
    \code{"R.squared"}, ad \code{"R"}
  }

\item{symmetric}{
    When no \code{y} argument is supplied this argument controls the format of
    the output band matrices. If \code{TRUE}, symmetric matrices are
    returned and, otherwise, an upper triangular matrices are returned
  }

\item{depth}{
    When \code{y} is not supplied, this parameter is mandatory and
    controls the maximum lag between columns of \code{x}
    considered. Thus, LD statistics are calculated between \code{x[,i]}
    and \code{x[,j]} only if \code{i} and \code{j} differ by no more
    than \code{depth}
  }

\item{LD_folder}{Locus-specific LD output folder.}
}
\description{
Get LD using \pkg{snpStats} package
}
\seealso{
Other LD: 
\code{\link{LD_1KG_download_vcf}()},
\code{\link{LD_1KG}()},
\code{\link{LD_blocks}()},
\code{\link{LD_ukbiobank}()},
\code{\link{calculate_LD}()},
\code{\link{construct_subset_vcf_name}()},
\code{\link{custom_panel}()},
\code{\link{dprime_table}()},
\code{\link{filter_LD}()},
\code{\link{filter_vcf_gaston}()},
\code{\link{filter_vcf}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{index_vcf}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{leadSNP_block}()},
\code{\link{load_or_create}()},
\code{\link{plink_LD}()},
\code{\link{plink_file}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{query_vcf}()},
\code{\link{rds_to_npz}()},
\code{\link{read_bin}()},
\code{\link{read_ld_table}()},
\code{\link{run_plink_LD}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_MAF}()},
\code{\link{translate_population}()},
\code{\link{vcf_to_bed}()}
}
\concept{LD}
\keyword{internal}
