% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpstats_get_MAF.R
\name{snpstats_get_MAF}
\alias{snpstats_get_MAF}
\title{Get MAF using \pkg{snpStats} package}
\source{
\href{https://www.bioconductor.org/packages/release/bioc/html/snpStats.html}{snpStats Bioconductor page}
}
\usage{
snpstats_get_MAF(dat, ss, force_new_MAF = FALSE, verbose = TRUE)
}
\arguments{
\item{LD_folder}{Locus-specific LD output folder.}
}
\description{
Get MAF using \pkg{snpStats} package
}
\seealso{
Other LD: 
\code{\link{LD_1KG_download_vcf}()},
\code{\link{LD_1KG}()},
\code{\link{LD_custom}()},
\code{\link{LD_ukbiobank}()},
\code{\link{compute_LD}()},
\code{\link{filter_LD}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{load_or_create}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{rds_to_npz}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{translate_population}()}
}
\concept{LD}
\keyword{internal}
