% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sparse.R
\name{to_sparse}
\alias{to_sparse}
\title{Convert to sparse}
\usage{
to_sparse(X, verbose = TRUE)
}
\arguments{
\item{X}{A \link[base]{matrix} / \link[base]{data.frame}.}

\item{verbose}{Print messages.}
}
\value{
Sparse matrix.
}
\description{
Convert a \link[base]{matrix} / \link[base]{data.frame} to a sparse matrix.
If \code{X} is already a sparse matrix, it simply returns \code{X} directly.
}
\examples{
mat <- to_sparse(X = mtcars)
}
