% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_population.R
\name{translate_population}
\alias{translate_population}
\title{Translate superopulation acronyms}
\usage{
translate_population(superpopulation)
}
\arguments{
\item{superpopulation}{Three-letter superpopulation name.}
}
\description{
Ensures a common ontology for synonynmous superpopulation names.
}
\seealso{
Other LD: 
\code{\link{LD_1KG_download_vcf}()},
\code{\link{LD_1KG}()},
\code{\link{LD_custom}()},
\code{\link{LD_ukbiobank}()},
\code{\link{compute_LD}()},
\code{\link{filter_LD}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{load_or_create}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{rds_to_npz}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_MAF}()}
}
\concept{LD}
\keyword{internal}
