% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryFunc.R
\name{tryFunc}
\alias{tryFunc}
\title{tryCatch extension}
\usage{
tryFunc(input, func, ...)
}
\arguments{
\item{input}{Function input.}

\item{func}{Function.}
}
\description{
Extension of tryCatch function.
}
\seealso{
Other general: 
\code{\link{get_os}()},
\code{\link{printer}()}
}
\concept{general}
\keyword{internal}
