% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_to_bed.R
\name{vcf_to_bed}
\alias{vcf_to_bed}
\title{Convert vcf file to BED file}
\usage{
vcf_to_bed(
  vcf.gz.subset,
  locus_dir,
  plink_prefix = "plink",
  conda_env = "echoR",
  verbose = TRUE
)
}
\arguments{
\item{vcf.gz.subset}{Path to the gzipped locus subset vcf.}

\item{locus_dir}{Locus-specific results directory.}
}
\description{
Uses plink to convert vcf to BED.
}
\seealso{
Other LD: 
\code{\link{LD_1KG_download_vcf}()},
\code{\link{LD_1KG}()},
\code{\link{LD_blocks}()},
\code{\link{LD_ukbiobank}()},
\code{\link{calculate_LD}()},
\code{\link{construct_subset_vcf_name}()},
\code{\link{custom_panel}()},
\code{\link{dprime_table}()},
\code{\link{filter_LD}()},
\code{\link{filter_vcf_gaston}()},
\code{\link{filter_vcf}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{index_vcf}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{leadSNP_block}()},
\code{\link{load_or_create}()},
\code{\link{plink_LD}()},
\code{\link{plink_file}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{query_vcf}()},
\code{\link{rds_to_npz}()},
\code{\link{read_bin}()},
\code{\link{read_ld_table}()},
\code{\link{run_plink_LD}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_LD}()},
\code{\link{snpstats_get_MAF}()},
\code{\link{translate_population}()}
}
\concept{LD}
\keyword{internal}
