% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_to_plink.R
\name{vcf_to_plink}
\alias{vcf_to_plink}
\title{Convert VCF to PLINK}
\usage{
vcf_to_plink(
  vcf,
  output_prefix = NULL,
  bcftools_path = NULL,
  make_bed = TRUE,
  recode = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{vcf}{Specify full name of .vcf or .vcf.gz file.}

\item{output_prefix}{Specify prefix for output files.}

\item{make_bed}{Create a new binary fileset. 
Unlike the automatic text-to-binary converters
(which only heed chromosome filters), this supports all of
PLINK's filtering flags.}

\item{recode}{Create a new text fileset with all filters applied.}
}
\description{
Convert a Variant Call File (VCF) to
 \href{https://www.cog-genomics.org/plink/}{PLINK} format.
}
\examples{
vcf <- system.file("extdata", "BST1.1KGphase3.vcf.bgz",
    package = "echodata")
paths <- vcf_to_plink(vcf = vcf)
}
