% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_LD_blocks.R
\name{compute_LD_blocks}
\alias{compute_LD_blocks}
\title{Compute LD blocks}
\source{
\href{https://github.com/pneuvial/adjclust}{adjclust GitHub}
}
\usage{
compute_LD_blocks(
  x,
  stats = c("R.squared", "D.prime"),
  type = c("capushe", "bstick"),
  k.max = NULL,
  pct = 0.15,
  verbose = TRUE
)
}
\arguments{
\item{x}{either a genotype matrix of class
\code{\link[snpStats:SnpMatrix-class]{SnpMatrix}}/\code{\link{matrix}} or a
linkage disequilibrium matrix of class
\code{\link[Matrix:dgCMatrix-class]{dgCMatrix}}. In the latter case the LD
values are expected to be in [0,1]}

\item{stats}{a character vector specifying the linkage disequilibrium
measures to be calculated (using the \code{\link[snpStats:ld]{ld}}
function) when \code{x} is a genotype matrix. Only "R.squared" and
"D.prime" are allowed, see Details.}

\item{type}{model selection approach between slope heuristic 
(\code{"capushe"}) and broken stick approach (\code{"bstick"})}

\item{k.max}{maximum number of clusters that can be selected. Default to 
\code{NULL}, in which case it is set to 
\eqn{\min(\max(100, \frac{n}{\log(n)}), \frac{n}{2})} where \eqn{n} is the
number of objects to be clustered for capushe and to \eqn{n} for the broken
stick model}

\item{pct}{minimum percentage of points for the plateau selection in 
capushe selection. See \code{\link[capushe]{DDSE}} for further details}

\item{verbose}{Print messages.}
}
\description{
Compute LD blocks using \link[adjclust]{snpClust}.
}
\keyword{internal}
