% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_UKB_LD.R
\name{download_UKB_LD}
\alias{download_UKB_LD}
\title{Download UKB LD}
\usage{
download_UKB_LD(
  LD.prefixes,
  locus_dir,
  alkes_url = paste("https://data.broadinstitute.org", "alkesgroup/UKBB_LD", sep = "/"),
  background = TRUE,
  force_overwrite = FALSE,
  download_method = "axel",
  conda_env = "echoR",
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{locus_dir}{Storage directory to use.}

\item{download_method}{\itemize{
\item{\code{"axel"} : }{Multi-threaded}
\item{\code{"wget"} : }{Single-threaded}
\item{\code{"download.file"} : }{Single-threaded}
\item{\code{"internal"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
\item{\code{"wininet"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
\item{\code{"libcurl"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
\item{\code{"curl"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
}}

\item{conda_env}{Conda environments to search in. 
If \code{NULL} (default), will search all conda environments.}

\item{nThread}{Number of threads to parallelize over.}

\item{verbose}{Print messages.}
}
\description{
Download UK Biobank Linkage Disequilibrium, 
pre-computed in 3Mb windows by the Alkes Price lab (Broad Institute).
}
\keyword{internal}
