% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_to_granges.R
\name{dt_to_granges}
\alias{dt_to_granges}
\title{Convert data.table to GRanges object}
\usage{
dt_to_granges(
  dat,
  style = c("NCBI", "UCSC"),
  chrom_col = "CHR",
  start_col = "POS",
  end_col = "POS"
)
}
\description{
Convert data.table to GRanges object
}
\keyword{internal}
