% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LD_blocks.R
\name{get_LD_blocks}
\alias{get_LD_blocks}
\title{Identify the LD block in which the lead SNP resides}
\source{
\href{https://github.com/pneuvial/adjclust}{adjclust GitHub}
}
\usage{
get_LD_blocks(
  query_dat,
  ss,
  stats = c("R.squared", "D.prime"),
  pct = 0.15,
  verbose = TRUE
)
}
\arguments{
\item{query_dat}{SNP-level data table.}

\item{ss}{\pkg{snpStats} object or LD matrix (containing r or r^2 values).}

\item{stats}{a character vector specifying the linkage disequilibrium
measures to be calculated (using the \code{\link[snpStats:ld]{ld}}
function) when \code{x} is a genotype matrix. Only "R.squared" and
"D.prime" are allowed, see Details.}

\item{pct}{minimum percentage of points for the plateau selection in 
capushe selection. See \code{\link[capushe]{DDSE}} for further details}

\item{verbose}{Print messages.}
}
\value{
A list with the input data and LD matrix (r^2),
}
\description{
Identify the LD block in which the lead SNP resides
}
