% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lead_r2.R
\name{get_lead_r2}
\alias{get_lead_r2}
\title{Get LD with lead SNP}
\usage{
get_lead_r2(
  dat,
  LD_matrix = NULL,
  fillNA = 0,
  LD_format = "matrix",
  verbose = TRUE
)
}
\arguments{
\item{dat}{SNP-level data.}

\item{LD_matrix}{LD matrix.}

\item{fillNA}{Value to fill NAs with in r/r2 columns.}

\item{LD_format}{The format of the provided \code{LD_matrix}:
"matrix" (wide) or "df" (long).}

\item{verbose}{Print messages.}
}
\value{
\link[data.table]{data.table} with the columns r and r2.
}
\description{
Add new columns r and r2 containing the degree of LD in each SNP (row)
 with the lead GWAS/QTL SNP.
}
\examples{
dat2 <- echoLD::get_lead_r2(dat = echodata::BST1,
                            LD_matrix = echodata::BST1_LD_matrix)
}
