% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_leadsnp_block.R
\name{get_leadsnp_block}
\alias{get_leadsnp_block}
\title{Get lead SNP block}
\source{
\href{https://github.com/pneuvial/adjclust}{adjclust GitHub}
}
\usage{
get_leadsnp_block(query_dat, ss, pct = 0.15, verbose = TRUE)
}
\arguments{
\item{verbose}{Print messages.}
}
\description{
Identify the LD block in which the lead SNP resides.
}
\keyword{internal}
