% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_UKB_MAF.R
\name{get_UKB_MAF}
\alias{get_UKB_MAF}
\title{Get MAF from UK Biobank.}
\source{
\href{http://biobank.ctsu.ox.ac.uk/showcase/field.cgi?id=22801}{UKB}
}
\usage{
get_UKB_MAF(
  dat,
  output_path = tools::R_user_dir(package = "echoLD", which = "cache"),
  force_new_maf = FALSE,
  download_method = "axel",
  nThread = 1,
  verbose = TRUE,
  conda_env = "echoR"
)
}
\arguments{
\item{dat}{SNP-level data.}

\item{output_path}{Path to store UKB_MAF file in.}

\item{force_new_maf}{Download UKB_MAF file again.}

\item{download_method}{\itemize{
\item{\code{"axel"} : }{Multi-threaded}
\item{\code{"wget"} : }{Single-threaded}
\item{\code{"download.file"} : }{Single-threaded}
\item{\code{"internal"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
\item{\code{"wininet"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
\item{\code{"libcurl"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
\item{\code{"curl"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
}
 or
\code{"download.file"} (single-threaded) .}

\item{nThread}{Number of threads to parallelize over.}

\item{verbose}{Print messages.}

\item{conda_env}{Conda environment to use.}
}
\description{
If \strong{MAF} column is missing,
download MAF from UK Biobank and use that instead.
}
\examples{
dat <- echodata::BST1
dat$MAF <- NULL
dat2 <- echoLD::get_UKB_MAF(dat = dat)
}
\concept{standardizing functions}
