% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD_1KG.R
\name{LD_1KG}
\alias{LD_1KG}
\title{Compute LD from 1000 Genomes}
\usage{
LD_1KG(
  locus_dir,
  dat,
  LD_reference = "1KGphase1",
  superpopulation = "EUR",
  vcf_folder = NULL,
  remote_LD = TRUE,
  LD_block = FALSE,
  LD_block_size = 0.7,
  remove_correlates = FALSE,
  remove_tmps = TRUE,
  fillNA = 0,
  download_method = "wget",
  as_sparse = TRUE,
  nThread = 1,
  conda_env = "echoR",
  verbose = TRUE
)
}
\arguments{
\item{LD_reference}{Which linkage disequilibrium reference panel do you want to use.
Options include:
\describe{
\item{"UKB"}{A pre-caclulated LD reference matrix from a subset of caucasian British individuals from the UK Biobank. See \href{https://www.biorxiv.org/content/10.1101/807792v2}{Wiessbrod et al. (2019)} for more details.}
\item{"1KGphase1"}{Download a subset of the 1000 Genomes Project Phase 1 vcf and calculate LD on the fly with plink.}
\item{"1KGphase3"}{Download a subset of the 1000 Genomes Project Phase 3 vcf and calculate LD on the fly with plink.}
\item{"<path>/*.vcf" or "<path>/*.vcf.gz"}{Alternatively, users can provide their own custom panel by supplying a list of \emph{.vcf} file path (one per locus) which \pkg{echolocatoR} will use to compute LD (using \emph{plink}).}
}}

\item{superpopulation}{Subset your LD reference panel by superopulation.
Setting the superpopulation is not currently possible when \code{LD_reference="UKB"}.
\href{https://www.internationalgenome.org/faq/which-populations-are-part-your-study/}{1KGphase1 options} include:
\describe{
\item{"AFR"}{African [descent]}
\item{"AMR"}{Ad-mixed American}
\item{"EAS"}{East Asian}
\item{"EUR"}{European}
\item{"SAS"}{South Asian}
}}

\item{remote_LD}{When acquiring LD matrices,
the default is to delete the full vcf or npz files after \pkg{echolocatoR} has extracted the necssary subset.
However, if you wish to keep these full files (which can be quite large) set \code{remote_LD=T}.}

\item{LD_block}{Calculate LD blocks with \emph{plink} and only include the block to which the lead SNP belongs.}

\item{LD_block_size}{Adjust the granularity of block sizes when \code{LD_block=T}.}

\item{remove_correlates}{A named list, where the names are the RSIDs of SNPs
whose LD correlates you wish to remove,
and the value is the absolute r2 threshold you wish to filter at for each RSID respectively
(e.g. \code{ remove_correlates = c("rs76904798"=.2, "rs10000737"=.8)}).
This will also remove the SNPs in \code{remove_correlates} themselves.}

\item{remove_tmps}{Whether to remove any temporary files (e.g. FINEMAP output files) after the pipeline is done running.}

\item{fillNA}{When pairwise LD (r) between two SNPs is \code{NA}, replace with 0.}

\item{conda_env}{The name of a conda environment to use.}

\item{verbose}{Whether \pkg{echolocatoR} should be verbose or silent.}
}
\description{
Downloads a subset vcf of the 1KG database that matches your locus coordinates.
Then uses \emph{plink} to calculate LD on the fly.
}
\details{
This approach is taken, because other API query tools have limitations with the window size being queried.
This approach does not have this limitations, allowing you to fine-map loci more completely.
}
\examples{
\dontrun{
data("BST1")
data("locus_dir")
BST1 <- limit_SNPs(max_snps = 500, dat = BST1)
LD_matrix <- LD_1KG(locus_dir = file.path("~/Desktop", locus_dir), dat = BST1, LD_reference = "1KGphase1")

## Kunkle et al 2019
locus_dir <- "/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Kunkle_2019/ACE"
}
}
\seealso{
Other LD: 
\code{\link{LD_1KG_download_vcf}()},
\code{\link{LD_blocks}()},
\code{\link{LD_ukbiobank}()},
\code{\link{calculate_LD}()},
\code{\link{construct_subset_vcf_name}()},
\code{\link{custom_panel}()},
\code{\link{dprime_table}()},
\code{\link{filter_LD}()},
\code{\link{filter_vcf_gaston}()},
\code{\link{filter_vcf}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{index_vcf}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{leadSNP_block}()},
\code{\link{load_or_create}()},
\code{\link{plink_LD}()},
\code{\link{plink_file}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{query_vcf}()},
\code{\link{rds_to_npz}()},
\code{\link{read_bin}()},
\code{\link{read_ld_table}()},
\code{\link{run_plink_LD}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_LD}()},
\code{\link{snpstats_get_MAF}()},
\code{\link{translate_population}()},
\code{\link{vcf_to_bed}()}
}
\concept{LD}
\keyword{internal}
