% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD_1KG_download_vcf.R
\name{LD_1KG_download_vcf}
\alias{LD_1KG_download_vcf}
\title{Download VCF subset from 1000 Genomes}
\usage{
LD_1KG_download_vcf(
  dat,
  LD_reference = "1KGphase1",
  remote_LD = TRUE,
  vcf_folder = NULL,
  locus_dir,
  locus = NULL,
  whole_vcf = FALSE,
  download_method = "axel",
  force_new_vcf = FALSE,
  query_by_regions = FALSE,
  remove_tmps = TRUE,
  nThread = 1,
  conda_env = "echoR",
  verbose = TRUE
)
}
\arguments{
\item{dat}{GWAS summary statistics subset to query the LD panel with.}

\item{LD_reference}{LD reference to use:
\itemize{
\item{"1KGphase1" : }{1000 Genomes Project Phase 1}
\item{"1KGphase3" : }{1000 Genomes Project Phase 3}
\item{"UKB" : }{Pre-computed LD from a British 
European-decent subset of UK Biobank.}
}}

\item{remote_LD}{Whether to pull the LD reference from remote repository, 
or locally stored files.}

\item{locus_dir}{Storage directory to use.}

\item{download_method}{\itemize{
\item{\code{"axel"} : }{Multi-threaded}
\item{\code{"wget"} : }{Single-threaded}
\item{\code{"download.file"} : }{Single-threaded}
\item{\code{"internal"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
\item{\code{"wininet"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
\item{\code{"libcurl"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
\item{\code{"curl"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
}
 or
\code{"download.file"} (single-threaded) .}

\item{query_by_regions}{You can make queries with \code{tabix}
in two different ways: 
\describe{
\item{\code{query_by_regions=F} \emph{(default)}}{
Return a vcf with all positions between the min/max in \code{dat}.
Takes up more storage but is MUCH faster}
\item{\code{query_by_regions=T}}{
Return a vcf with only the exact positions present in \code{dat}.
 Takes up less storage but is MUCH slower}
}}

\item{remove_tmps}{Remove all temporary files like VCF, npz, and plink files.}

\item{nThread}{Number of threads to parallelize over.}

\item{conda_env}{Conda environment name.}

\item{verbose}{Print messages.}
}
\description{
Query the 1000 Genomes Project for a subset of their individual-level VCF
files.
}
\details{
\code{
data("BST1")
vcf_subset.popDat <- LD_1KG_download_vcf(
    dat = BST1,
    LD_reference = "1KGphase1",
    locus_dir = file.path(tempdir(), locus_dir)
)
}
}
\seealso{
Other LD: 
\code{\link{LD_1KG}()},
\code{\link{LD_blocks}()},
\code{\link{LD_ukbiobank}()},
\code{\link{calculate_LD}()},
\code{\link{construct_subset_vcf_name}()},
\code{\link{custom_panel}()},
\code{\link{dprime_table}()},
\code{\link{filter_LD}()},
\code{\link{filter_vcf_gaston}()},
\code{\link{filter_vcf}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{index_vcf}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{leadSNP_block}()},
\code{\link{load_or_create}()},
\code{\link{plink_LD}()},
\code{\link{plink_file}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{query_vcf}()},
\code{\link{rds_to_npz}()},
\code{\link{read_bin}()},
\code{\link{read_ld_table}()},
\code{\link{run_plink_LD}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_LD}()},
\code{\link{snpstats_get_MAF}()},
\code{\link{translate_population}()},
\code{\link{vcf_to_bed}()}
}
\concept{LD}
\keyword{internal}
