% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD_1KG_download_vcf.R
\name{LD_1KG_download_vcf}
\alias{LD_1KG_download_vcf}
\title{Download VCF subset from 1000 Genomes}
\usage{
LD_1KG_download_vcf(
  dat,
  LD_reference = "1KGphase1",
  superpopulation = NULL,
  samples = NULL,
  local_storage = NULL,
  locus_dir,
  force_new_vcf = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dat}{GWAS summary statistics subset to query the LD panel with.}

\item{LD_reference}{LD reference to use:
\itemize{
\item{"1KGphase1" : }{1000 Genomes Project Phase 1}
\item{"1KGphase3" : }{1000 Genomes Project Phase 3}
\item{"UKB" : }{Pre-computed LD from a British
European-decent subset of UK Biobank.}
}}

\item{superpopulation}{Superpopulation to subset LD panel by
(used only if \code{LD_reference} is "1KGphase1" or "1KGphase3".)}

\item{samples}{Sample names to subset the VCF by before computing LD.}

\item{local_storage}{Storage folder for previously downloaded LD files.
If \code{LD_reference} is "1KGphase1" or "1KGphase3",
\code{local_storage} is where VCF files are stored.
If \code{LD_reference} is "UKB", \code{local_storage} is where
LD compressed numpy array (npz) files are stored.
Set to \code{NULL} to download VCFs/LD npz from remote storage system.}

\item{locus_dir}{Storage directory to use.}

\item{force_new_vcf}{Force the creation of a new LD file even if one exists.}

\item{verbose}{Print messages.}
}
\description{
Query the 1000 Genomes Project for a subset of their individual-level VCF
files.
}
\details{
\code{
data("BST1")
vcf_subset.popDat <- LD_1KG_download_vcf(
    dat = BST1,
    LD_reference = "1KGphase1",
    locus_dir = file.path(tempdir(), locus_dir)
)
}
}
\seealso{
Other LD: 
\code{\link{LD_1KG}()},
\code{\link{LD_custom}()},
\code{\link{LD_ukbiobank}()},
\code{\link{compute_LD}()},
\code{\link{filter_LD}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{load_or_create}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{rds_to_npz}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_MAF}()},
\code{\link{translate_population}()}
}
\concept{LD}
\keyword{internal}
