% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldlinkr_ldproxy_batch.R
\name{ldlinkr_ldproxy_batch}
\alias{ldlinkr_ldproxy_batch}
\title{Extract LD proxies from 1KGphase3}
\source{
\href{https://www.rdocumentation.org/packages/LDlinkR/versions/1.0.2}{
website}
}
\usage{
ldlinkr_ldproxy_batch(
  snp,
  pop = "CEU",
  r2d = "r2",
  min_corr = FALSE,
  save_dir = NULL,
  verbose = TRUE
)
}
\arguments{
\item{snp}{a character string or data frame listing rsID's or chromosome coordinates (e.g. "chr7:24966446"), one per line}

\item{pop}{a 1000 Genomes Project population, (e.g. YRI or CEU), multiple allowed, default = "CEU"}

\item{r2d}{either "r2" for LD R2 or "d" for LD D', default = "r2"}

\item{min_corr}{Minimum correlation with \code{snp}.}

\item{save_dir}{Save folder.}

\item{verbose}{Print messages.}
}
\description{
Wrapper for \code{LDlinkR::LDproxy_batch}.
Easy to use but doesn't scale up well to many SNPs (takes way too long).
}
\details{
\code{
data("merged_DT")
lead.snps <- setNames(subset(merged_DT, leadSNP)$Locus, 
                      subset(merged_DT, leadSNP)$SNP)
proxies <- ldlinkr_ldproxy_batch(snp = lead.snps)
}
}
\seealso{
Other LD: 
\code{\link{LD_1KG_download_vcf}()},
\code{\link{LD_1KG}()},
\code{\link{LD_blocks}()},
\code{\link{LD_ukbiobank}()},
\code{\link{calculate_LD}()},
\code{\link{construct_subset_vcf_name}()},
\code{\link{custom_panel}()},
\code{\link{dprime_table}()},
\code{\link{filter_LD}()},
\code{\link{filter_vcf_gaston}()},
\code{\link{filter_vcf}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{index_vcf}()},
\code{\link{leadSNP_block}()},
\code{\link{load_or_create}()},
\code{\link{plink_LD}()},
\code{\link{plink_file}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{query_vcf}()},
\code{\link{rds_to_npz}()},
\code{\link{read_bin}()},
\code{\link{read_ld_table}()},
\code{\link{run_plink_LD}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_LD}()},
\code{\link{snpstats_get_MAF}()},
\code{\link{translate_population}()},
\code{\link{vcf_to_bed}()}
}
\concept{LD}
\keyword{internal}
