% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_LD.R
\name{plot_LD}
\alias{plot_LD}
\title{Plot a subset of the LD matrix}
\usage{
plot_LD(LD_matrix, dat, span = 10, method = c("gaston", "heatmap", "image"))
}
\arguments{
\item{span}{This is very computationally intensive,
so you need to limit the number of SNPs with span.
If \code{span=10}, only 10 SNPs upstream and 10 SNPs downstream of the lead SNP will be plotted.}
}
\description{
Uses \code{gaston} to plot a SNP-annotated LD matrix.
}
\examples{
\dontrun{
data("BST1")
data("BST1_LD")
plot_LD(LD_matrix = LD_matrix, dat = BST1)
}
}
\seealso{
Other LD: 
\code{\link{LD_1KG_download_vcf}()},
\code{\link{LD_1KG}()},
\code{\link{LD_custom}()},
\code{\link{LD_ukbiobank}()},
\code{\link{compute_LD}()},
\code{\link{filter_LD}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{load_or_create}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{rds_to_npz}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_MAF}()},
\code{\link{translate_population}()}
}
\concept{LD}
\keyword{internal}
