% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_LD_list.R
\name{read_LD_list}
\alias{read_LD_list}
\title{Read LD list}
\usage{
read_LD_list(RDS_path, query_dat, verbose = TRUE)
}
\arguments{
\item{RDS_path}{Path to saved LD matrix.}

\item{query_dat}{Query data.frame.}

\item{verbose}{Print messages.}
}
\value{
LD list with 3 elements.
}
\description{
Import existing LD results.
}
\keyword{internal}
