% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_plink_LD.R
\name{run_plink_LD}
\alias{run_plink_LD}
\title{Calculate LD (r or r2)}
\usage{
run_plink_LD(
  bim,
  LD_folder,
  plink_prefix = "plink",
  r_format = "r",
  extract_file = NULL
)
}
\arguments{
\item{bim}{A bim file produced by \emph{plink}}

\item{LD_folder}{Locus-specific LD output folder.}

\item{r_format}{Whether to fill the matrix with \code{r} or \code{r2}.}
}
\description{
This appriach computes and LD matrix of r or r2 (instead of D') from a vcf.
See \code{dprime_table} for a slower (but more flexible) alternative to computing LD.
}
\examples{
\dontrun{
data("LRRK2")
LD_folder <- "/Users/schilder/Desktop/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/LRRK2/plink/saved"
bim_path <- file.path(LD_folder, "plink.bim")
bim <- data.table::fread(bim_path, col.names = c("CHR", "SNP", "V3", "POS", "A1", "A2"), stringsAsFactors = F)
bim <- subset(bim, SNP \%in\% LRRK2$SNP)
ld.bin <- file.path(LD_folder, paste0("plink", ".ld.bin"))
SNPs <- data.table::fread(file.path(LD_folder, "SNPs.txt"), col.names = "RSID")
bin.vector <- readBin(ld.bin, what = "numeric", n = length(SNPs$RSID)^2)
}
}
\seealso{
Other LD: 
\code{\link{LD_1KG_download_vcf}()},
\code{\link{LD_1KG}()},
\code{\link{LD_blocks}()},
\code{\link{LD_ukbiobank}()},
\code{\link{calculate_LD}()},
\code{\link{construct_subset_vcf_name}()},
\code{\link{custom_panel}()},
\code{\link{dprime_table}()},
\code{\link{filter_LD}()},
\code{\link{filter_vcf_gaston}()},
\code{\link{filter_vcf}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{index_vcf}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{leadSNP_block}()},
\code{\link{load_or_create}()},
\code{\link{plink_LD}()},
\code{\link{plink_file}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{query_vcf}()},
\code{\link{rds_to_npz}()},
\code{\link{read_bin}()},
\code{\link{read_ld_table}()},
\code{\link{saveSparse}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_LD}()},
\code{\link{snpstats_get_MAF}()},
\code{\link{translate_population}()},
\code{\link{vcf_to_bed}()}
}
\concept{LD}
\keyword{internal}
