% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_LD_matrix.R
\name{save_LD_matrix}
\alias{save_LD_matrix}
\title{Save LD matrix}
\usage{
save_LD_matrix(
  LD_matrix,
  dat,
  locus_dir,
  fillNA = 0,
  LD_reference,
  subset_common = TRUE,
  as_sparse = TRUE,
  verbose = TRUE
)
}
\description{
Save an LD matrix after initial pre-processing.
}
\details{
\code{
#data("BST1")
#data("BST1_LD_matrix")
#data("locus_dir")
#locus_dir <- file.path(tempdir(),locus_dir)
#LD_list <- save_LD_matrix(
#    LD_matrix = BST1_LD_matrix,
#    dat = BST1,
#    locus_dir = locus_dir,
#    LD_reference = "UKB")

#LD_list <- save_LD_matrix(
#    LD_matrix = BST1_LD_matrix,
#    dat = BST1,
#    locus_dir = locus_dir,
#    LD_reference = "custom_vcf")
}
}
\seealso{
Other LD: 
\code{\link{LD_1KG_download_vcf}()},
\code{\link{LD_1KG}()},
\code{\link{LD_blocks_cli}()},
\code{\link{LD_custom_cli}()},
\code{\link{LD_custom}()},
\code{\link{LD_ukbiobank}()},
\code{\link{calculate_LD}()},
\code{\link{compute_LD}()},
\code{\link{construct_subset_vcf_name}()},
\code{\link{dprime_table}()},
\code{\link{filter_LD}()},
\code{\link{filter_vcf_cli}()},
\code{\link{filter_vcf_gaston}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{index_vcf_cli}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{leadsnp_block_plink}()},
\code{\link{load_or_create}()},
\code{\link{plink_LD}()},
\code{\link{plink_file}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{query_vcf_cli}()},
\code{\link{query_vcf}()},
\code{\link{rds_to_npz}()},
\code{\link{read_bin}()},
\code{\link{read_ld_table}()},
\code{\link{run_plink_LD}()},
\code{\link{saveSparse}()},
\code{\link{snpstats_get_MAF}()},
\code{\link{translate_population}()},
\code{\link{vcf_to_bed_cli}()}
}
\concept{LD}
\keyword{internal}
