% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveSparse.R
\name{saveSparse}
\alias{saveSparse}
\title{Save LD matrix as a sparse matrix}
\usage{
saveSparse(LD_matrix, LD_path, verbose = TRUE)
}
\description{
Converting LD matrices to sparse format reduces file size by half.
}
\seealso{
Other LD: 
\code{\link{LD_1KG_download_vcf}()},
\code{\link{LD_1KG}()},
\code{\link{LD_custom}()},
\code{\link{LD_ukbiobank}()},
\code{\link{check_population_1kg}()},
\code{\link{compute_LD}()},
\code{\link{filter_LD}()},
\code{\link{get_locus_vcf_folder}()},
\code{\link{ldlinkr_ldproxy_batch}()},
\code{\link{load_or_create}()},
\code{\link{plot_LD}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{rds_to_npz}()},
\code{\link{save_LD_matrix}()},
\code{\link{snpstats_get_MAF}()}
}
\concept{LD}
\keyword{internal}
