% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_vcf_samples.R
\name{select_vcf_samples}
\alias{select_vcf_samples}
\title{Subset VCF samples}
\usage{
select_vcf_samples(
  superpopulation = NULL,
  samples = NULL,
  LD_reference = "1KGphase1",
  verbose = TRUE
)
}
\arguments{
\item{samples}{[Optional] Sample names to subset the VCF by. 
If this option is used, the \link[GenomicRanges]{GRanges} object will be
converted to a \link[VariantAnnotation]{ScanVcfParam} for usage by 
\link[VariantAnnotation]{readVcf}.}

\item{verbose}{Print messages.}
}
\description{
Subsrt VCF samples by superpopulation.
}
