% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_common_snps.R
\name{subset_common_snps}
\alias{subset_common_snps}
\title{Subset LD matrix and dataframe to only their shared SNPs}
\usage{
subset_common_snps(LD_matrix, dat, fillNA = 0, verbose = FALSE)
}
\arguments{
\item{LD_matrix}{LD matrix.}

\item{dat}{SNP-level summary statistics subset 
to query the LD panel with.}

\item{fillNA}{Value to fill LD matrix NAs with.}

\item{verbose}{Print messages.}
}
\value{
data.frame
}
\description{
Find the SNPs that are shared between an LD matrix and another data.frame 
with a "SNP" column. Then remove any non-shared SNPs from both objects.
}
\examples{
out <- echoLD::subset_common_snps(LD_matrix = echodata::BST1_LD_matrix,
                                  dat = echodata::BST1)
}
\concept{SNP filters}
